+++ services/__init__.py
@@ -5,6 +5,7 @@ from .branding import BrandingService
 from .session_service import SessionService
 from .stats_service import StatsService
 from .api_service import APIService, api_service
+from .event_service import EventService, event_service
 
 __all__ = [
     "ConfigService",
@@ -13,4 +14,6 @@ __all__ = [
     "StatsService",
     "APIService",
     "api_service",
+    "EventService",
+    "event_service",
 ]